[evals] Separate the lm-eval adapter from the finestore contract
finestore.eval defined the shared evaluation schema and also knew how to read
lm-eval's --log_samples rows. The conversion belongs with the mechanism, so it
moves to marin.evaluation.lm_eval_samples along with the export and rebuild that
drive it. finestore keeps the models, the archive tables, the row round-trip,
EvaluationStore, and the metric/filter ranking a reader applies to stored rows.
evaldash still imports only finestore.

The export now preserves every artifact the harness left in the results tree
rather than the two filename patterns it reads. run_artifacts lists the tree
instead of globbing it, which is what reaches evalchemy's .resume state: a `*`
glob skips dot-directories, so fingerprint.json and manifest.jsonl were the one
thing a rebuilt archive could not account for. It excludes the archive's own
tables and markers, which share the run's root, so a re-export cannot fold the
archive into itself.

Fleet-wide archive sweeps are administrative, so backfill-samples and
rebuild-samples move to experiments.evaluation.migrations.cli, next to
migrate_archive. experiments.evaluation.cli is now only the launcher.
